# docs/openapi.yaml
openapi: 3.1.0
info:
  title: PetMon API
  description: Pet monitoring and nutrition tracking API
  version: 1.0.0
  contact:
    name: PetMon Team
  license:
    name: MIT
servers:
  - url: /api/v1
    description: Main API server
security:
  - bearerAuth: []
paths:
...
  /elimination/records:
    get:
      tags:
        - Elimination Records
      summary: List elimination records
      x-required-scope: api_read
      description: Retrieve elimination records with optional filtering
      parameters:
        - name: pet_id
          in: query
          required: false
          schema: { type: string, format: uuid }
        - name: date
          in: query
          required: false
          schema: { type: string, format: date }
        - name: date_from
          in: query
          required: false
          schema: { type: string, format: date }
        - name: date_to
          in: query
          required: false
          schema: { type: string, format: date }
        - name: event_type
          in: query
          required: false
          schema: { $ref: '#/components/schemas/EliminationEventType' }
        - name: limit
          in: query
          required: false
          schema: { type: integer, format: int64, minimum: 1 }
        - name: offset
          in: query
          required: false
          schema: { type: integer, format: int64, minimum: 0 }
      responses:
        '200':
          description: List of elimination records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EliminationRecord'
        '403':
          $ref: '#/components/responses/Forbidden'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Elimination Records
      summary: Create elimination record
      x-required-scope: api_write
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEliminationRecord'
      responses:
        '201':
          description: Record created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EliminationRecord'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /elimination/records/{id}:
    get:
      tags:
        - Elimination Records
      summary: Get elimination record by ID
      x-required-scope: api_read
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Elimination record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EliminationRecord'
        '403':
          $ref: '#/components/responses/Forbidden'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags:
        - Elimination Records
      summary: Update elimination record
      x-required-scope: api_write
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEliminationRecord'
      responses:
        '200':
          description: Record updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EliminationRecord'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Elimination Records
      summary: Delete elimination record
      x-required-scope: api_write
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '204':
          description: Record deleted
        '403':
          $ref: '#/components/responses/Forbidden'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /nutrition/records:
    get:
      tags:
        - Nutrition Records
      summary: List nutrition records
      x-required-scope: api_read
      description: Retrieve nutrition records with optional filtering
      parameters:
        - name: pet_id
          in: query
          required: false
          description: Filter by pet UUID
          schema:
            type: string
            format: uuid
        - name: date
          in: query
          required: false
          description: Filter by specific date (YYYY-MM-DD)
          schema:
            type: string
            format: date
        - name: date_from
          in: query
          required: false
          description: Filter records from this date (YYYY-MM-DD, inclusive)
          schema:
            type: string
            format: date
        - name: date_to
          in: query
          required: false
          description: Filter records until this date (YYYY-MM-DD, inclusive)
          schema:
            type: string
            format: date
        - name: category
          in: query
          required: false
          description: Filter by nutrition category
          schema:
            $ref: '#/components/schemas/NutritionCategory'
        - name: limit
          in: query
          required: false
          description: Maximum number of records to return
          schema:
            type: integer
            format: int64
            minimum: 1
        - name: offset
          in: query
          required: false
          description: Pagination offset
          schema:
            type: integer
            format: int64
            minimum: 0
      responses:
        '200':
          description: List of nutrition records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NutritionRecord'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Nutrition Records
      summary: Create nutrition record
      x-required-scope: api_write
      description: Create a single nutrition record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateNutritionRecord'
      responses:
        '201':
          description: Nutrition record created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NutritionRecord'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /nutrition/records/{id}:
    get:
      tags:
        - Nutrition Records
      summary: Get nutrition record by ID
      x-required-scope: api_read
      description: Retrieve a specific nutrition record
      parameters:
        - name: id
          in: path
          required: true
          description: Record ID
          schema:
            type: string
      responses:
        '200':
          description: Nutrition record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NutritionRecord'
        '403':
          $ref: '#/components/responses/Forbidden'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags:
        - Nutrition Records
      summary: Update nutrition record
      x-required-scope: api_write
      description: Update one or more fields of a nutrition record
      parameters:
        - name: id
          in: path
          required: true
          description: Record ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateNutritionRecord'
      responses:
        '200':
          description: Nutrition record updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NutritionRecord'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Nutrition Records
      summary: Delete nutrition record
      x-required-scope: api_write
      description: Delete a nutrition record
      parameters:
        - name: id
          in: path
          required: true
          description: Record ID
          schema:
            type: string
      responses:
        '204':
          description: Nutrition record deleted
        '403':
          $ref: '#/components/responses/Forbidden'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Bearer authentication with either:
        - OIDC JWT token (format: standard JWT) — full access, no scope restrictions
        - API token with `pm_api_` prefix — access controlled by the token's scope set

        **API token scopes:**
        | Scope | Access granted |
        |---|---|
        | `all` | Full access — equivalent to OIDC |
        | `api_read` | All GET endpoints |
        | `api_write` | All POST / PATCH / DELETE endpoints |
        | `mcp` | MCP tool endpoint (`/mcp`) |

        A token may hold multiple scopes (e.g. `["api_read", "mcp"]`).
        The scope required by each endpoint is shown in the `x-required-scope` field.
  responses:
    BadRequest:
      description: Invalid request parameters or body
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication required or invalid credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Token does not have the required scope for this endpoint
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
      required:
        - error
...
    NutritionCategory:
      type: string
      enum:
        - wet_food
        - dry_food
        - water
        - liquids
      description: Category of nutrition intake
    NutritionRecord:
      type: object
      properties:
        id:
          type: string
          description: Record ID
        pet_id:
          type: string
          format: uuid
          description: Pet UUID
        occurred_at:
          type: string
          format: date-time
          description: When the intake occurred (RFC3339)
        local_date:
          type: string
          format: date
          description: Local date of the occurrence
        category:
          $ref: '#/components/schemas/NutritionCategory'
        amount:
          type: number
          format: double
          description: Amount consumed
        unit:
          type: string
          nullable: true
          description: Unit of measurement (ml, g, etc.)
        note:
          type: string
          nullable: true
          description: Optional note about this intake (food type, medication, etc.)
        source_type:
          type: string
          description: Source of the record (e.g., manual, telegram)
        created_at:
          type: string
          format: date-time
          description: Record creation timestamp (RFC3339)
        updated_at:
          type: string
          format: date-time
          description: Record last update timestamp (RFC3339)
      required:
        - id
        - pet_id
        - occurred_at
        - local_date
        - category
        - amount
        - source_type
        - created_at
        - updated_at
    CreateNutritionRecord:
      type: object
      properties:
        pet_id:
          type: string
          format: uuid
          description: Pet UUID
        occurred_at:
          type: string
          format: date-time
          description: When the intake occurred (RFC3339)
        local_date:
          type: string
          format: date
          nullable: true
          description: Local date (auto-derived if not provided)
        category:
          $ref: '#/components/schemas/NutritionCategory'
        amount:
          type: number
          format: double
          description: Amount consumed
        unit:
          type: string
          nullable: true
          description: Unit of measurement
        note:
          type: string
          nullable: true
          description: Optional note about this intake
        source_type:
          type: string
          nullable: true
          description: Source of the record (defaults to manual)
      required:
        - pet_id
        - occurred_at
        - category
        - amount
    UpdateNutritionRecord:
      type: object
      properties:
        occurred_at:
          type: string
          format: date-time
          nullable: true
        local_date:
          type: string
          format: date
          nullable: true
        category:
          $ref: '#/components/schemas/NutritionCategory'
        amount:
          type: number
          format: double
          nullable: true
        unit:
          type: string
          nullable: true
        note:
          type: string
          nullable: true
...
    EliminationEventType:
      type: string
      enum:
        - general
        - urination
        - defecation
        - vomit
        - no_output
      description: "Type of elimination event. Subtypes for defecation: normal|soft|liquid|hard|blood|mucus. Subtypes for vomit: food|fur|fur_with_food|bile|other. no_output = pet went to the toilet but produced nothing."
    EliminationRecord:
      type: object
      properties:
        id:
          type: string
        pet_id:
          type: string
          format: uuid
        occurred_at:
          type: string
          format: date-time
          description: Naive local datetime (RFC3339, no Z suffix)
        local_date:
          type: string
          format: date
        event_type:
          $ref: '#/components/schemas/EliminationEventType'
        subtype:
          type: string
          nullable: true
          description: Event-specific subtype (defecation or vomit only)
        duration_seconds:
          type: integer
          nullable: true
          description: Time spent in seconds (optional)
        note:
          type: string
          nullable: true
        source_type:
          type: string
        is_auto_categorized:
          type: boolean
          description: True when event_type was set by auto-categorization; false for manual entries and edits.
        auto_categorize_confidence:
          type: number
          format: float
          nullable: true
          description: Classifier confidence (0–1) when auto-tagged by the contextual model; null for manual entries, legacy bucket match, or after user correction.
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - pet_id
        - occurred_at
        - local_date
        - event_type
        - source_type
        - is_auto_categorized
        - auto_categorize_confidence
        - created_at
        - updated_at
    CreateEliminationRecord:
      type: object
      properties:
        pet_id:
          type: string
          format: uuid
        event_type:
          $ref: '#/components/schemas/EliminationEventType'
        occurred_at:
          type: string
          format: date-time
          nullable: true
          description: >
            Naive local datetime (YYYY-MM-DDTHH:MM:SS). Omit for a real-time
            entry (server stamps now). When omitted with local_date, the
            time-of-day is now on that journal day.
        local_date:
          type: string
          format: date
          nullable: true
          description: >
            Journal day. Defaults to the date of occurred_at, or today when
            occurred_at is also omitted.
        subtype:
          type: string
          nullable: true
        duration_seconds:
          type: integer
          nullable: true
        note:
          type: string
          nullable: true
        source_type:
          type: string
          nullable: true
      required:
        - pet_id
        - event_type
...
    UpdateEliminationRecord:
      type: object
      properties:
        event_type:
          $ref: '#/components/schemas/EliminationEventType'
        occurred_at:
          type: string
          format: date-time
          nullable: true
        local_date:
          type: string
          format: date
          nullable: true
        subtype:
          type: string
          nullable: true
        duration_seconds:
          type: integer
          nullable: true
        note:
          type: string
          nullable: true
...